Move query processor implementation classes into separate files
